Specified buckets for Prometheus histograms

+++ internal/server/workload/workload_driver.go
@@ -1197,7 +1197,7 @@ func (d *BasicWorkloadDriver) handleEvent(evt domain.Event) error {
 			sessionLifetimeDuration := time.Since(session.GetCreatedAt())
 			metrics.PrometheusMetricsWrapperInstance.WorkloadSessionLifetimeSeconds.
 				With(prometheus.Labels{"workload_id": d.workload.GetId()}).
-				Observe(sessionLifetimeDuration.Seconds())
+				Observe(float64(sessionLifetimeDuration.Seconds()))
 		}
 
 		d.workload.SessionStopped(traceSessionId)
